fonts, changed containers to components, added in description functionality
made it so you can add descriptions, added in a new font, changed it to have container components for the 1st 2nd and 3rd level views and added in some nice details in terms of user functionality. Need to make it so that scrollview is used for adding, editing and last three views. And maybe max 2 line titles.